Use size_t for sizes.
	* include/urbi/uvalue.hh (UValue::size): here.
	* src/libuco/uvalue-common.cc: and there.

+++ sdk-remote/include/urbi/uvalue.hh
@@ -57,10 +57,12 @@ namespace urbi
     {
       return *array[i+offset];
     }
-    int size() const
+
+    size_t size() const
     {
-      return static_cast<int>(array.size());
+      return array.size();
     }
+
     void setOffset(int n)
     {
       offset = n;
